[hackday] Separating code with packages :w.error:
